requirements.txt
fastapi
uvicorn
pillow
torch
opencv-python
numpy
huggingface
transformers
python-multipart
pandas
soundfile
meteomatics
requests
